Relocate source and refactor/clarify modules
